Create a "forms" package to avoid a too long forms.py file
